Populated and tested MoneyList class
