+++ mt.log
@@ -53,3 +53,21 @@ Mar 19, 2025 8:39:36 PM seedu.duke.MTLogger logInfo
 INFO: Loading previous entries from mt.txt
 Mar 19, 2025 8:41:30 PM seedu.duke.MTLogger logInfo
 INFO: Loading previous entries from mt.txt
+Mar 20, 2025 3:11:37 PM seedu.duke.MTLogger logInfo
+INFO: Loading previous entries from mt.txt
+Mar 20, 2025 3:12:03 PM seedu.duke.MTLogger logWarning
+WARNING: Invalid index provided: 4
+Mar 20, 2025 3:12:04 PM seedu.duke.MTLogger logInfo
+INFO: Deleted entry at index: 1
+Mar 20, 2025 3:15:18 PM seedu.duke.MTLogger logInfo
+INFO: Loading previous entries from mt.txt
+Mar 20, 2025 3:15:33 PM seedu.duke.MTLogger logInfo
+INFO: Loading previous entries from mt.txt
+Mar 20, 2025 3:16:42 PM seedu.duke.MTLogger logInfo
+INFO: Saving entries into mt.txt
+Mar 20, 2025 3:17:31 PM seedu.duke.MTLogger logInfo
+INFO: Loading previous entries from mt.txt
+Mar 20, 2025 3:17:50 PM seedu.duke.MTLogger logInfo
+INFO: Saving entries into mt.txt
+Mar 20, 2025 3:18:53 PM seedu.duke.MTLogger logInfo
+INFO: Loading previous entries from mt.txt
