+++ mt.log.1
@@ -110,3 +110,39 @@ Mar 19, 2025 8:41:45 PM seedu.duke.MTLogger logInfo
 INFO: Total expense calculated: 101.1
 Mar 19, 2025 8:41:55 PM seedu.duke.MTLogger logInfo
 INFO: Exiting CLI program.
+Mar 20, 2025 3:11:37 PM seedu.duke.MTLogger logInfo
+INFO: Starting CLI program.
+Mar 20, 2025 3:11:37 PM seedu.duke.MTLogger logInfo
+INFO: Loaded 5 entries from file.
+Mar 20, 2025 3:12:04 PM seedu.duke.MTLogger logInfo
+INFO: Saving entries into mt.txt
+Mar 20, 2025 3:15:18 PM seedu.duke.MTLogger logInfo
+INFO: Starting CLI program.
+Mar 20, 2025 3:15:18 PM seedu.duke.MTLogger logInfo
+INFO: Loaded 2 entries from file.
+Mar 20, 2025 3:15:33 PM seedu.duke.MTLogger logInfo
+INFO: Starting CLI program.
+Mar 20, 2025 3:15:33 PM seedu.duke.MTLogger logInfo
+INFO: Loaded 2 entries from file.
+Mar 20, 2025 3:16:42 PM seedu.duke.MTLogger logInfo
+INFO: Added expense: [Expense] foodstuff Value=$20.000000 
+Mar 20, 2025 3:16:53 PM seedu.duke.MTLogger logInfo
+INFO: Total expense calculated: 20.0
+Mar 20, 2025 3:17:20 PM seedu.duke.MTLogger logInfo
+INFO: Exiting CLI program.
+Mar 20, 2025 3:17:31 PM seedu.duke.MTLogger logInfo
+INFO: Starting CLI program.
+Mar 20, 2025 3:17:31 PM seedu.duke.MTLogger logInfo
+INFO: Loaded 3 entries from file.
+Mar 20, 2025 3:17:42 PM seedu.duke.MTLogger logInfo
+INFO: Total budget set to: 100.0
+Mar 20, 2025 3:17:50 PM seedu.duke.MTLogger logInfo
+INFO: Deleted entry at index: 0
+Mar 20, 2025 3:18:06 PM seedu.duke.MTLogger logInfo
+INFO: Exiting CLI program.
+Mar 20, 2025 3:18:53 PM seedu.duke.MTLogger logInfo
+INFO: Starting CLI program.
+Mar 20, 2025 3:18:53 PM seedu.duke.MTLogger logInfo
+INFO: Loaded 2 entries from file.
+Mar 20, 2025 3:19:04 PM seedu.duke.MTLogger logInfo
+INFO: Exiting CLI program.
