+++ mt.log.2
@@ -10,3 +10,5 @@ Mar 16, 2025 2:16:13 PM seedu.duke.MTLogger logInfo
 INFO: Saving entries into mt.txt
 Mar 18, 2025 3:56:25 PM seedu.duke.MTLogger logInfo
 INFO: Loading previous entries from mt.txt
+Mar 20, 2025 3:12:04 PM seedu.duke.MTLogger logInfo
+INFO: Deleted entry at index: 1
