+++ mt.log.3
@@ -12,3 +12,5 @@ Mar 18, 2025 3:56:25 PM seedu.duke.MTLogger logInfo
 INFO: Starting CLI program.
 Mar 18, 2025 3:56:25 PM seedu.duke.MTLogger logInfo
 INFO: Loaded 2 entries from file.
+Mar 20, 2025 3:12:04 PM seedu.duke.MTLogger logInfo
+INFO: Saving entries into mt.txt
